Move perlin.h to common/
